#580 Extracts reusable portion of the tiqr authentication and uses it in the existing UseApp page.
