+++ myconext-gui/src/constants/authenticationStatus.js
@@ -1,5 +1,6 @@
 export const authenticationStatus = {
     PENDING: "PENDING",
     SUSPENDED: "SUSPENDED",
-    SUCCESS: "SUCCESS"
+    SUCCESS: "SUCCESS",
+    TIMEOUT: "TIMEOUT"
 }
